Change Model data

+++ app/src/main/java/com/sharif/tech_challenge/data/model/SubMainModel.kt
@@ -2,7 +2,7 @@ package com.sharif.tech_challenge.data.model
 
 //region Card
 data class CardResult (
-        val code: Long,
+        val code: Int,
         val title: String="",
         val description: String="",
         val tag: String ="",
